metadata fetch now working in parameterstore
