+++ pkg/provider/aws/util/provider.go
@@ -14,8 +14,10 @@ limitations under the License.
 package util
 
 import (
+	"encoding/json"
 	"fmt"
 
+	awssm "github.com/aws/aws-sdk-go/service/secretsmanager"
 	esv1beta1 "github.com/external-secrets/external-secrets/apis/externalsecrets/v1beta1"
 )
 
@@ -58,4 +60,18 @@ func IsReferentSpec(prov esv1beta1.AWSAuth) bool {
 	}
 
 	return false
+}
+
+func TagsToJSONString(tags []*awssm.Tag) (string, error) {
+	tagMap := make(map[string]string, len(tags))
+	for _, tag := range tags {
+		tagMap[*tag.Key] = *tag.Value
+	}
+
+	byteArr, err := json.Marshal(tagMap)
+	if err != nil {
+		return "", err
+	}
+
+	return string(byteArr), nil
 }
